refactoring to guilds. remove "tag" parameter. extract service class
