+++ bot/__init__.py
@@ -1,3 +1,4 @@
 from .TS3Bot import Bot
+from .guild_service import GuildService
 
-__all__ = ['Bot']
+__all__ = ['GuildService', 'Bot']
