+++ bot/rest/server.py
@@ -3,10 +3,11 @@ import threading
 
 import flask_cors
 import waitress  # productive serve
-from flask import Flask, jsonify
+from flask import Flask
 from werkzeug.exceptions import HTTPException
 
 from bot.rest.controller import GuildController, HealthController, OtherController, ResetRosterController
+from bot.rest.utils import error_response
 
 LOG = logging.getLogger(__name__)
 
@@ -48,7 +49,7 @@ def create_http_server(bot, port=8080):
 def register_controllers(app, bot):
     controller = [
         HealthController(),
-        GuildController(bot),
+        GuildController(bot.guild_service),
         ResetRosterController(bot),
         OtherController(bot),
     ]
@@ -59,4 +60,4 @@ def register_controllers(app, bot):
 def register_error_handlers(flask: Flask):
     @flask.errorhandler(HTTPException)
     def _handle_error(e: HTTPException):
-        return jsonify(code=e.code, name=e.name, desc=e.description), e.code
+        return error_response(e.code, e.name, e.description)
