+++ bot/rest/utils/__init__.py
@@ -1,3 +1,4 @@
+from .error_utils import error_response
 from .utils import try_get
 
-__all__ = ['try_get']
+__all__ = ['try_get', 'error_response']
